requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
